fix dequant + minor refactor
